Shannon-Weaver Diversity Index

Compute the Shannon-Weaver Diversity Index (\(H'\)), Maximum diversity (\(H'_{max}\)) and Shannon Equitability Index (\(E_{H}\)) (Shannon and Weaver 1949) to compare the phenotypic diversity for qualitative traits between entire collection (EC) and core set (CS).

shannon.evaluate.core(data, names, qualitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

qualitative
Name of columns with the qualitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

A data frame with the following columns.

Trait

The qualitative trait.

EC_H

The Shannon-Weaver Diversity Index (\(H'\)) for EC.

EC_H

The Shannon-Weaver Diversity Index (\(H'\)) for CS.

EC_Hmax

The Maximum diversity value (\(H'_{max}\)) for EC.

CS_Hmax

The Maximum diversity value (\(H'_{max}\)) for CS.

EC_EH

The Shannon Equitability Index (\(E_{H}\)) for EC.

CS_EH

The Shannon Equitability Index (\(E_{H}\)) for CS.

Details

Shannon-Weaver Diversity Index (\(H'\)) is computed as follows.

\[H' = -\sum_{i=1}^{k}p_{i}log(p_{i})\]

Where \(p_{i}\) denotes the proportion in the group \(k\).

The maximum value of the index (\(H'_{max}\)) is \(\ln(k)\). This value occurs when each group has the same frequency.

The Shannon equitability index (\(E_{H}\)) is the Shannon diversity index divided by the maximum diversity.

\[E_{H} = \frac{H}{\ln{(k)}}\]

References

Shannon CE, Weaver W (1949). The Mathematical Theory of Communication, number v. 2 in The Mathematical Theory of Communication. University of Illinois Press.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

shannon.evaluate.core(data = ec, names = "genotypes",
                      qualitative = qual, selected = core)

#>   Trait      EC_H      CS_H  EC_Hmax  CS_Hmax     EC_EH     CS_EH
#> 1    X1 1.2089465 1.2310001 1.386294 1.386294 0.8720706 0.8879789
#> 2    X6 0.7799636 0.8501612 1.386294 1.386294 0.5626248 0.6132616
#> 3    X7 0.8830332 1.0566659 1.386294 1.386294 0.6369738 0.7622233