Chi-squared Test for Homogenity

Compare the distribution frequencies of qualitative traits between entire collection (EC) and core set (CS) by Chi-squared test for homogenity (Pearson 1900; Snedecor and Irwin 1933) .

chisquare.evaluate.core(data, names, qualitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

qualitative
Name of columns with the qualitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

A a data frame with the following columns.

Trait

The qualitative trait.

EC_No.Classes

The number of classes in the trait for EC.

EC_Classes

The frequency of the classes in the trait for EC.

CS_No.Classes

The number of classes in the trait for CS.

CS_Classes

The frequency of the classes in the trait for CS.

chisq_statistic

The χ2 test statistic.

chisq_pvalue

The p value for the test statistic.

chisq_significance

The significance of the test statistic (*: p ≤ 0.01; **: p ≤ 0.05; ns: p > 0.05).

References

Pearson K (1900). “X. On the criterion that a given system of deviations from the probable in the case of a correlated system of variables is such that it can be reasonably supposed to have arisen from random sampling.” The London, Edinburgh, and Dublin Philosophical Magazine and Journal of Science, 50(302), 157--175. bibtex[publisher=Taylor \& Francis] Citation Key: pearson\_ x.\_1900.

Snedecor G, Irwin M (1933). “On the chi-square test for homogeneity.” Iowa State College Journal of Science, 8, 75--81.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

chisquare.evaluate.core(data = ec, names = "genotypes",
                        qualitative = qual, selected = core)

#>   Trait EC_No.Classes                     EC_Classes CS_No.Classes
#> 1    X1             4 1(312); 2(915); 3(589); 4(155)             4
#> 2    X6             4  1(18); 2(901); 3(1035); 4(17)             4
#> 3    X7             4  1(64); 2(1183); 3(670); 4(54)             4
#>                   CS_Classes chisq_statistic chisq_pvalue chisq_significance
#> 1 1(37); 2(95); 3(46); 4(19)        4.411622   0.21867813                 ns
#> 2   1(7); 2(95); 3(94); 4(1)       12.133658   0.00979902                 **
#> 3  1(8); 2(88); 3(85); 4(16)       27.933293   0.00029997                 **