Diversity Indices

Compute the following diversity indices and perform corresponding statistical tests to compare the phenotypic diversity for qualitative traits between entire collection (EC) and core set (CS).

Simpson's and related indices

Simpson's Index (d) simpson_measurement_1949,peet_measurement_1974EvaluateCore

Simpson's Index of Diversity or Gini's Diversity Index or Gini-Simpson Index or Nei's Diversity Index or Nei's Variation Index (D) gini_variabilita_1912,gini_variabilita_1912-2,greenberg_measurement_1956,berger_diversity_1970,nei_analysis_1973,peet_measurement_1974EvaluateCore

Maximum Simpson's Index of Diversity or Maximum Nei's Diversity/Variation Index (D_max) hennink_interpretation_1990EvaluateCore

Simpson's Reciprocal Index or Hill's N_2 (D_R) williams_patterns_1964,hill_diversity_1973EvaluateCore

Relative Simpson's Index of Diversity or Relative Nei's Diversity/Variation Index (D') hennink_interpretation_1990EvaluateCore

Shannon-Weaver and related indices

Shannon or Shannon-Weaver or Shannon-Weiner Diversity Index (H) shannon_mathematical_1949,peet_measurement_1974EvaluateCore

Maximum Shannon-Weaver Diversity Index (H_max) hennink_interpretation_1990EvaluateCore

Relative Shannon-Weaver Diversity Index or Shannon Equitability Index (H') hennink_interpretation_1990EvaluateCore

McIntosh Diversity Index

McIntosh Diversity Index (D_Mc) mcintosh_index_1967,peet_measurement_1974EvaluateCore

diversity.evaluate.core(data, names, qualitative, selected, base = 2, R = 1000)

Arguments

data

The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names

Name of column with the individual names as a character string.

qualitative

Name of columns with the qualitative traits as a character vector.

selected

Character vector with the names of individuals selected in core collection and present in the names column.

base

The logarithm base to be used for computation of Shannon-Weaver Diversity Index (I). Default is 2.

R

The number of bootstrap replicates. Default is 1000.

Value

A list with three data frames as follows.

simpson

EC_No.Classes

The number of classes in the trait for EC.

CS_No.Classes

The number of classes in the trait for CS.

EC_d

The Simpson's Index (d) for EC.

EC_D

The Simpson's Index of Diversity (D) for EC.

EC_D.max

The Maximum Simpson's Index of Diversity (D_max) for EC.

EC_D.inv

The Simpson's Reciprocal Index (D_R) for EC.

EC_D.rel

The Relative Reciprocal Index (D') for EC.

EC_d.V

The variance of d for EC according to simpson_measurement_1949EvaluateCore.

EC_d.boot.V

The bootstrap variance of d for EC.

CS_d

The Simpson's Index (d) for CS.

CS_D

The Simpson's Index of Diversity (D) for CS.

CS_D.max

The Maximum Simpson's Index of Diversity (D_max) for CS.

CS_D.inv

The Simpson's Reciprocal Index (D_R) for CS.

CS_D.rel

The Relative Reciprocal Index (D') for CS.

CS_d.V

The variance of d for CS according to simpson_measurement_1949EvaluateCore.

CS_d.boot.V

The bootstrap variance of d for CS.

d.t.df

The degrees of freedom for t test.

d.t.stat

The t statistic.

d.t.pvalue

The p value for t test.

d.t.significance

The significance of t test for t-test

d.boot.z.df

The degrees of freedom for bootstrap z score.

d.boot.z.stat

The bootstrap z score.

d.boot.z.pvalue

The p value of z score.

d.boot.z.significance

The significance of z score.

shannon

EC_No.Classes

The number of classes in the trait for EC.

CS_No.Classes

The number of classes in the trait for CS.

EC_I

The Shannon-Weaver Diversity Index (I) for EC.

EC_I.max

The Maximum Shannon-Weaver Diversity Index (I_max) for EC.

EC_I.rel

The Relative Shannon-Weaver Diversity Index (I') for EC.

EC_I.V

The variance of I for EC according to hutcheson_test_1970EvaluateCore.

EC_I.boot.V

The bootstrap variance of I for EC.

CS_I

The Shannon-Weaver Diversity Index (I) for CS.

CS_I.max

The Maximum Shannon-Weaver Diversity Index (I_max) for CS.

CS_I.rel

The Relative Shannon-Weaver Diversity Index (I') for CS.

CS_I.V

The variance of I for CS according to hutcheson_test_1970EvaluateCore.

CS_I.boot.V

The bootstrap variance of I for CS.

I.t.stat

The t statistic.

I.t.df

The degrees of freedom for t test.

I.t.pvalue

The p value for t test.

I.t.significance

The significance of t test for t-test

I.boot.z.df

The degrees of freedom for bootstrap z score.

I.boot.z.stat

The bootstrap z score.

I.boot.z.pvalue

The p value of z score.

I.boot.z.significance

The significance of z score.

mcintosh

CS_No.Classes

The number of classes in the trait for CS.

EC_D.Mc

The McIntosh Index (D_Mc) for EC.

CS_D.Mc

The McIntosh Index (D_Mc) for CS.

M.boot.z.stat

The bootstrap z score.

M.boot.z.df

The degrees of freedom for bootstrap z score.

M.boot.z.pvalue

The p value of z score.

M.boot.z.significance

The significance of z score.

Details

The diversity indices and the corresponding statistical tests implemented in diversity.evaluate.core are as follows.

Simpson's and related indices

Simpson's index (d) which estimates the probability that two accessions randomly selected will belong to the same phenotypic class of a trait, is computed as follows simpson_measurement_1949,peet_measurement_1974EvaluateCore.

d = _i = 1^kp_i^2

Where, p_i denotes the proportion/fraction/frequency of accessions in the ith phenotypic class for a trait and k is the number of phenotypic classes for the trait.

The value of d can range from 0 to 1 with 0 representing maximum diversity and 1, no diversity.

d is subtracted from 1 to give Simpson's index of diversity (D) greenberg_measurement_1956,berger_diversity_1970,peet_measurement_1974,hennink_interpretation_1990EvaluateCore originally suggested by gini_variabilita_1912,gini_variabilita_1912-2;textualEvaluateCore and described in literature as Gini's diversity index or Gini-Simpson index. It is the same as Nei's diversity index or Nei's variation index nei_analysis_1973,hennink_interpretation_1990EvaluateCore. Greater the value of D, greater the diversity with a range from 0 to 1.

D = 1 - d

The maximum value of D, D_max occurs when accessions are uniformly distributed across the phenotypic classes and is computed as follows hennink_interpretation_1990EvaluateCore.

D_max = 1 - 1k

Reciprocal of d gives the Simpson's reciprocal index (D_R) williams_patterns_1964,hennink_interpretation_1990EvaluateCore and can range from 1 to k. This was also described in hill_diversity_1973;textualEvaluateCore as (N_2).

D_R = 1d

Relative Simpson's index of diversity or Relative Nei's diversity/variation index (H') hennink_interpretation_1990EvaluateCore is defined as follows peet_measurement_1974EvaluateCore.

D' = DD_max

Differences in Simpson's diversity index for qualitative traits of EC and CS can be tested by a t-test using the associated variance estimate described in simpson_measurement_1949;textualEvaluateCore lyons_c20_1978EvaluateCore.

The t statistic is computed as follows.

t = d_EC - d_CSV_d_EC + V_d_CS

Where, the variance of d (V_d) is,

V_d = 4N(N-1)(N-2)_i=1^k(p_i)^3 + 2N(N-1)_i=1^k(p_i)^2 - 2N(N-1)(2N-3) ( _i=1^k(p_i)^2 )^2[N(N-1)]^2

The associated degrees of freedom is computed as follows.

df = (k_EC - 1) + (k_CS - 1)

Where, k_EC and k_CS are the number of phenotypic classes in the trait for EC and CS respectively.

Shannon-Weaver and related indices

An index of information H, was described by shannon_mathematical_1949;textualEvaluateCore as follows.

H = -_i=1^kp_i _2(p_i)

H is described as Shannon or Shannon-Weaver or Shannon-Weiner diversity index in literature.

Alternatively, H is also computed using natural logarithm instead of logarithm to base 2.

H = -_i=1^kp_i (p_i)

The maximum value of H (H_max) is (k). This value occurs when each phenotypic class for a trait has the same proportion of accessions.

H_max = _2(k)\;\; OR \;\; H_max = (k)

The relative Shannon-Weaver diversity index or Shannon equitability index (H') is the Shannon diversity index (I) divided by the maximum diversity (H_max).

H' = HH_max

Differences in Shannon-Weaver diversity index for qualitative traits of EC and CS can be tested by Hutcheson t-test hutcheson_test_1970EvaluateCore.

The Hutcheson t statistic is computed as follows.

t = H_EC - H_CSV_H_EC + V_H_CS

Where, the variance of H (V_H) is,

V_H = _i=1^kn_i(_2n_i)^2 (_i=1^k_2n_i)^2NN^2

OR

V_H = _i=1^kn_i(n_i)^2 (_i=1^kn_i)^2NN^2

The associated degrees of freedom is approximated as follows.

df = (V_H_EC + V_H_CS)^2V_H_EC^2N_EC + V_H_CS^2N_CS

McIntosh Diversity Index

A similar index of diversity was described by mcintosh_index_1967;textualEvaluateCore as follows (D_Mc) peet_measurement_1974EvaluateCore.

D_Mc = N - _i=1^kn_i^2N - N

Where, n_i denotes the number of accessions in the ith phenotypic class for a trait and N is the total number of accessions so that p_i = n_i/N.

Testing for difference with bootstrapping

Bootstrap statistics are employed to test the difference between the Simpson, Shannon-Weaver and McIntosh indices for qualitative traits of EC and CS solow_simple_1993EvaluateCore.

If I_EC and I_CS are the diversity indices with the original number of accessions, then random samples of the same size as the original are repeatedly generated (with replacement) R times and the corresponding diversity index is computed for each sample.

I_EC^* = H_EC_1, H_EC_, , H_EC_R

I_CS^* = H_CS_1, H_CS_, , H_CS_R

Then the bootstrap null sample I_0 is computed as follows.

^* = I_EC^* - I_CS^*

I_0 = ^* - ^*

Where, ^* is the mean of ^*.

Now the original difference in diversity indices (_0 = I_EC - I_CS) is tested against mean of bootstrap null sample (I_0) by a z test. The z score test statistic is computed as follows.

z = _0 - H_0V_H_0

Where, H_0 and V_H_0 are the mean and variance of the bootstrap null sample H_0.

The corresponding degrees of freedom is estimated as follows.

df = (k_EC - 1) + (k_CS - 1)

References

See also

Examples

data("cassava_CC")
data("cassava_EC")

ec <- cbind(genotypes = rownames(cassava_EC), cassava_EC)
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL

core <- rownames(cassava_CC)

quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

ec[, qual] <- lapply(ec[, qual],
                     function(x) factor(as.factor(x)))

# \donttest{
diversity.evaluate.core(data = ec, names = "genotypes",
                        qualitative = qual, selected = core)
#> $simpson
#>     Trait EC_No.Classes CS_No.Classes              EC_d              EC_D
#> 1    CUAL             5             4 0.379231526565524 0.620768473434476
#> 2    LNGS             3             3 0.422400150078142 0.577599849921858
#> 3    PTLC             5             5 0.477398006104682 0.522601993895318
#> 4    DSTA             5             5 0.351400071089646 0.648599928910354
#> 5    LFRT             5             4 0.448935347916114 0.551064652083886
#> 6   LBTEF             6             6 0.223971174841036 0.776028825158964
#> 7    CBTR             3             3  0.51874284166756  0.48125715833244
#> 8    NMLB            10             9 0.206703725435988 0.793296274564012
#> 9    ANGB             4             4 0.350516387291879  0.64948361270812
#> 10 CUAL9M             5             5 0.309073521363567 0.690926478636433
#> 11  LVC9M             5             5 0.425299451029954 0.574700548970046
#> 12 TNPR9M             5             5 0.247165582455527 0.752834417544473
#> 13   PL9M             3             2 0.498635332682619 0.501364667317381
#> 14   STRP             4             4 0.322189843207836 0.677810156792164
#> 15   STRC             2             2 0.510325630074306 0.489674369925694
#> 16   PSTR             3             2 0.555306757465823 0.444693242534177
#>             EC_D.max         EC_D.inv          EC_D.rel               EC_d.V
#> 1                0.8 2.63691156971154 0.775960591793095  6.2401326584027e-05
#> 2  0.666666666666667 2.36742340128195 0.866399774882787 2.66303075243519e-05
#> 3                0.8 2.09468826265002 0.653252492369147 6.43256463380354e-05
#> 4                0.8 2.84575924216273 0.810749911137942 4.49202770089222e-05
#> 5                0.8 2.22749223165839 0.688830815104857 3.64033438958825e-05
#> 6  0.833333333333333 4.46486026922774 0.931234590190757 1.01110395238633e-05
#> 7  0.666666666666667 1.92773744459853  0.72188573749866 4.63756492257553e-05
#> 8                0.9 4.83784217188518 0.881440305071124 1.57355754103247e-05
#> 9               0.75 2.85293366089411 0.865978150277494 3.10366375721298e-05
#> 10               0.8 3.23547612745411 0.863658098295541 1.34905804296262e-05
#> 11               0.8 2.35128448338761 0.718375686212558 4.98247003940663e-05
#> 12               0.8 4.04587074812462 0.941043021930592 2.56650302792209e-05
#> 13 0.666666666666667   2.005473608579 0.752047000976072 3.79031161219323e-06
#> 14              0.75 3.10376016215672 0.903746875722886 9.34647324671549e-06
#> 15               0.5  1.9595331707216 0.979348739851389 1.21791747807758e-05
#> 16 0.666666666666667 1.80080646697613 0.667039863801265  6.1357244865428e-05
#>             EC_d.boot.V              CS_d              CS_D          CS_D.max
#> 1  6.68533450338314e-05 0.378472222222222 0.621527777777778              0.75
#> 2  2.65879113857189e-05 0.387755102040816 0.612244897959184 0.666666666666667
#> 3  6.50363094034534e-05 0.421343537414966 0.578656462585034               0.8
#> 4  4.79746735208086e-05 0.303500566893424 0.696499433106576               0.8
#> 5  3.49356184489191e-05 0.426587301587302 0.573412698412698              0.75
#> 6  9.78416887814862e-06 0.200538548752834 0.799461451247165 0.833333333333333
#> 7  4.76384743057889e-05 0.487244897959184 0.512755102040816 0.666666666666667
#> 8  1.55424381336112e-05 0.198058390022676 0.801941609977324 0.888888888888889
#> 9  3.19616815160651e-05 0.342120181405896 0.657879818594104              0.75
#> 10 1.35390135457135e-05 0.288052721088435 0.711947278911565               0.8
#> 11 4.73810044068571e-05 0.388676303854875 0.611323696145125               0.8
#> 12 2.63678232813366e-05   0.2218679138322   0.7781320861678               0.8
#> 13 3.95700024252201e-06 0.500283446712018 0.499716553287982               0.5
#> 14 9.00891406453171e-06 0.311862244897959 0.688137755102041              0.75
#> 15 1.11259330666434e-05 0.518140589569161 0.481859410430839               0.5
#> 16 5.79882678523648e-05 0.555555555555556 0.444444444444444               0.5
#>            CS_D.inv          CS_D.rel               CS_d.V          CS_d.boot.V
#> 1  2.64220183486239 0.828703703703704 0.000664974207441761 0.000711409006695385
#> 2  2.57894736842105 0.918367346938775 0.000253494320337725 0.000240277035602111
#> 3  2.37336024217962 0.723320578231292  0.00106773305088108  0.00104909427501709
#> 4  3.29488676161569  0.87062429138322 0.000464316667250201   0.0004640806330644
#> 5  2.34418604651163 0.764550264550265 0.000434218289438062 0.000428639235160984
#> 6  4.98657243816254 0.959353741496599 6.05627470284696e-05 6.05634678928432e-05
#> 7  2.05235602094241 0.769132653061224 0.000257136051752956 0.000271330248841789
#> 8  5.04901610017889  0.90218431122449 0.000216132859442877 0.000204816497715932
#> 9  2.92294946147473 0.877173091458806 0.000360254817691854   0.0003537099748542
#> 10 3.47158671586716 0.889934098639456 0.000144703074224106 0.000152475008559194
#> 11 2.57283500455789 0.764154620181406 0.000503057233986995 0.000488621895795101
#> 12 4.50718620249122 0.972665107709751 0.000134121919501297 0.000143869065469617
#> 13 1.99886685552408 0.999433106575964 2.11737525161229e-05  1.4057939053243e-05
#> 14 3.20654396728016 0.917517006802721 8.69838142388246e-05 8.64975280349916e-05
#> 15 1.92997811816193 0.963718820861678  0.00022467592663519 0.000201749293623147
#> 16              1.8 0.888888888888889 0.000601970661850899 0.000601085481209453
#>    d.t.df             d.t.stat         d.t.pvalue d.t.significance d.boot.z.df
#> 1       7     0.02815376878219  0.978325351754391               ns           7
#> 2       4     2.06997700767549   0.10723294813912               ns           4
#> 3       8     1.66600295615057  0.134274950280504               ns           8
#> 4       8     2.12261420536943 0.0665538249488766               ns           8
#> 5       7      1.0301574783912  0.337209790540494               ns           7
#> 6      10     2.78735177789005 0.0192068432217533                *          10
#> 7       4     1.80798358527422  0.144886591272101               ns           4
#> 8      17    0.567755084270463  0.577623407084813               ns          17
#> 9       6    0.424456215383704  0.686031514328123               ns           6
#> 10      8      1.6713011659847  0.133206249679844               ns           8
#> 11      8     1.55754090990284  0.157957033254601               ns           8
#> 12      8     2.00128917457502 0.0803555866511265               ns           8
#> 13      3   -0.329859967245677  0.763196306139146               ns           3
#> 14      6     1.05224751693921  0.333206278005286               ns           6
#> 15      2   -0.507791421657648  0.662061546648583               ns           2
#> 16      3 -0.00966012858117729   0.99289893656246               ns           3
#>          d.boot.z.stat    d.boot.z.pvalue d.boot.z.significance
#> 1   0.0276014468259671   0.97875044140037                    ns
#> 2     2.11134701701347   0.10233493680879                    ns
#> 3     1.67600791343304  0.132263448642767                    ns
#> 4     2.12071979935888 0.0667503673247065                    ns
#> 5     1.03756609320658  0.333978621629479                    ns
#> 6     2.77065034924187 0.0197652293840094                     *
#> 7     1.76028327597653  0.153166166195353                    ns
#> 8    0.577008613944405  0.571496330602201                    ns
#> 9    0.425217529507766  0.685506175360378                    ns
#> 10    1.63791870182241  0.140072980275729                    ns
#> 11    1.58556233614025  0.151499007860614                    ns
#> 12    1.93374355649653 0.0892037747747804                    ns
#> 13    0.39244722767569   0.72094875991824                    ns
#> 14    1.03972126692463  0.338541348748645                    ns
#> 15    0.54146070930643   0.64244084391876                    ns
#> 16 0.00959298293692684     0.992948292625                    ns
#> 
#> $shannon
#>     Trait EC_No.Classes CS_No.Classes              EC_I          EC_I.max
#> 1    CUAL             5             4  1.60118025300384   1.6094379124341
#> 2    LNGS             3             3  1.34638092645118  1.09861228866811
#> 3    PTLC             5             5  1.29757991678528   1.6094379124341
#> 4    DSTA             5             5  1.75217791456181   1.6094379124341
#> 5    LFRT             5             4  1.32020430332539   1.6094379124341
#> 6   LBTEF             6             6  2.26107917725546  1.79175946922805
#> 7    CBTR             3             3  1.04940762729587  1.09861228866811
#> 8    NMLB            10             9  2.46043656290491  2.30258509299405
#> 9    ANGB             4             4  1.63181278826841  1.38629436111989
#> 10 CUAL9M             5             5  1.81895733334165   1.6094379124341
#> 11  LVC9M             5             5  1.45781261313429   1.6094379124341
#> 12 TNPR9M             5             5  2.15936383275215   1.6094379124341
#> 13   PL9M             3             2  1.02393231699601  1.09861228866811
#> 14   STRP             4             4  1.70471635517281  1.38629436111989
#> 15   STRC             2             2 0.985051563686417 0.693147180559945
#> 16   PSTR             3             2 0.932902628052214  1.09861228866811
#>             EC_I.rel               EC_I.V          EC_I.boot.V
#> 1  0.994869227718287 0.000495550546246209 0.000483808510194458
#> 2   1.22552873323805  0.00026683714279372 0.000269948802236365
#> 3  0.806231732681647  0.00076741293709653 0.000765752404583755
#> 4   1.08868934988106 0.000644551217715297 0.000699283275380162
#> 5     0.820289054412   0.0005480962422548 0.000517842713945963
#> 6   1.26193231630002 0.000259759685726472 0.000271033385039332
#> 7  0.955211987086096 0.000312696844542415 0.000303798121154971
#> 8   1.06855402234261 0.000496178249738945 0.000505626109031182
#> 9   1.17710410864701 0.000306008701151613  0.00030810953251426
#> 10  1.13018173567856 0.000397228605602063 0.000382025104334203
#> 11  0.90578990458197 0.000698631874341118 0.000665418478253527
#> 12  1.34168818571345 0.000264448274894998 0.000279879360986154
#> 13 0.932023360340673 0.000100041535943681 0.000101884272024935
#> 14  1.22969291586506  0.00021912035039574 0.000213652548557445
#> 15  1.42112900595031 2.53468656861686e-05 2.75676958917147e-05
#> 16 0.849164566676391 0.000206905486943159 0.000215549640442331
#>                 CS_I          CS_I.max          CS_I.rel               CS_I.V
#> 1   1.60715392433211  1.38629436111989  1.15931649828959  0.00466900952090494
#> 2   1.44881563572518  1.09861228866811  1.31876882378736  0.00183098114336171
#> 3   1.60685189446012   1.6094379124341 0.998393216691366  0.00915269666156106
#> 4   1.94652505102181   1.6094379124341  1.20944401519528  0.00501290672204481
#> 5   1.39390699808299  1.38629436111989  1.00549135679738  0.00481441616742965
#> 6    2.3831722498942  1.79175946922805    1.330073757568  0.00172264312694065
#> 7   1.12338192727968  1.09861228866811  1.02254629669362  0.00290287113955504
#> 8   2.56080478714959  2.19722457733622  1.16547248449867  0.00546163553548553
#> 9   1.69700114892879  1.38629436111989  1.22412757097122  0.00354899161566695
#> 10  1.92688046691671   1.6094379124341  1.19723814881589  0.00380714696318465
#> 11  1.59702066650884   1.6094379124341 0.992284731315619  0.00676974620846264
#> 12  2.24730159981114   1.6094379124341  1.39632699245436  0.00129552549169971
#> 13 0.999591034189098 0.693147180559945  1.44210502794168 7.02197242512667e-06
#> 14  1.74920954558179  1.38629436111989  1.26178796844324   0.0020863467552808
#> 15  0.97366806454962 0.693147180559945  1.40470608819769 0.000443961107972778
#> 16  0.91829583405449 0.693147180559945  1.32482084585941  0.00132275132275134
#>             CS_I.boot.V           I.t.stat           I.t.df          I.t.pvalue
#> 1    0.0047929123394758 -0.083123660050637 205.323489838581   0.933834195011031
#> 2   0.00187875640207067  -2.23647066374941 220.068588438007  0.0263241034578903
#> 3   0.00931911218476163  -3.10514822627186 197.214845421711 0.00218182272218803
#> 4   0.00517797969455639  -2.58385107753543 213.627424864319  0.0104369229742795
#> 5   0.00488761731418891  -1.00646624014807  208.16009186019   0.315359849464886
#> 6   0.00187862560896072  -2.74217439229004 221.982338302534 0.00660128212793374
#> 7   0.00298070544341396  -1.30452382793487  205.90491319267   0.193511253079244
#> 8   0.00547694938695042   -1.3003275918761 199.746997945695   0.194987041994507
#> 9   0.00362924040839815  -1.04992396937323 198.073409342984   0.295032389672131
#> 10  0.00434669657011922  -1.66442333074281 204.664065671499   0.097558069731458
#> 11  0.00669021208688097  -1.61083557606938 204.247107694347   0.108759859193219
#> 12   0.0013166833981556  -2.22646968440541 242.577447747539  0.0269016983557465
#> 13 4.24533303750304e-05   2.35246259571922 1837.93308927511  0.0187545059227561
#> 14  0.00212088460832853 -0.926646485957522 204.916304701679     0.3552001943646
#> 15 0.000438689236849775   0.52546865211132 187.669671190779   0.599877733009115
#> 16  0.00146139768748281  0.373471875118422 224.120762900574   0.709150296735105
#>    I.t.significance I.boot.z.df      I.boot.z.stat    I.boot.z.pvalue
#> 1                ns           7 0.0817336348245371  0.937146464242981
#> 2                 *           4    2.2064267784542 0.0919914613051522
#> 3                **           8   3.08665470359158 0.0149652109934245
#> 4                 *           8   2.57117798516249 0.0330669187074517
#> 5                ns           7   1.01631020857175  0.343315003404297
#> 6                **          10   2.66359626201418 0.0237522092011334
#> 7                ns           4   1.28540612042092  0.268024804385408
#> 8                ns          17    1.2958056339159  0.212360842372221
#> 9                ns           6   1.04598381571325  0.335865437056982
#> 10               ns           8    1.5855588008592  0.151499807281787
#> 11               ns           8   1.60319190075875  0.147559909191939
#> 12                *           8   2.20200971458306 0.0588093099330856
#> 13                *           3   1.98106680901202  0.141910185247116
#> 14               ns           6  0.935780231278676  0.385503284430038
#> 15               ns           2  0.532654452382285  0.647528455859605
#> 16               ns           3  0.355800409404779  0.745538558892587
#>    I.boot.z.significance
#> 1                     ns
#> 2                     ns
#> 3                      *
#> 4                      *
#> 5                     ns
#> 6                      *
#> 7                     ns
#> 8                     ns
#> 9                     ns
#> 10                    ns
#> 11                    ns
#> 12                    ns
#> 13                    ns
#> 14                    ns
#> 15                    ns
#> 16                    ns
#> 
#> $mcintosh
#>     Trait EC_No.Classes CS_No.Classes           EC_D.Mc           CS_D.Mc
#> 1    CUAL             5             4 0.393778012434178 0.416968887221027
#> 2    LNGS             3             3 0.358820738546382  0.40884307929476
#> 3    PTLC             5             5 0.316779500161785 0.380225207834671
#> 4    DSTA             5             5 0.417380876241608 0.486635921075666
#> 5    LFRT             5             4 0.338215427137732 0.375861861492538
#> 6   LBTEF             6             6 0.539900644138108 0.598348264293068
#> 7    CBTR             3             3 0.286749615427161 0.327215963767637
#> 8    NMLB            10             9 0.558974562677424 0.601358283540837
#> 9    ANGB             4             4 0.418145333938704 0.449791785009108
#> 10 CUAL9M             5             5 0.455147453452433 0.502026774187731
#> 11  LVC9M             5             5 0.356538441668976 0.408042032313971
#> 12 TNPR9M             5             5 0.515402077468806 0.573194332783244
#> 13   PL9M             3             2 0.301198604544138 0.317162433070916
#> 14   STRP             4             4 0.443181993303366 0.478468357680793
#> 15   STRC             2             2 0.292763414762376 0.303603702882552
#> 16   PSTR             3             2 0.261175379595452 0.275932675604997
#>        M.boot.z.stat M.boot.z.df      M.boot.z.pvalue M.boot.z.significance
#> 1   1.00391168633292           7    0.348854575801086                    ns
#> 2   3.39520672163674           4   0.0273980949692247                     *
#> 3   2.28323260882762           8    0.051809250272037                    ns
#> 4    3.3110316235721           8   0.0106817822611812                     *
#> 5   2.15395666678097           7   0.0682139277106039                    ns
#> 6   5.96821850593738          10 0.000137837562706689                    **
#> 7   2.99754729409007           4   0.0400387096312927                     *
#> 8   2.42994156651831          17   0.0264706618143546                     *
#> 9   1.80568141934553           6     0.12099659678039                    ns
#> 10   3.8867793107122           8  0.00463023359089327                    **
#> 11  2.53855378664644           8   0.0347890152183959                     *
#> 12  4.24645671583821           8  0.00281266593084537                    **
#> 13  4.42355476586344           3   0.0214543888693413                     *
#> 14  3.81879071089912           6  0.00877307594548063                    **
#> 15 0.913484492624288           2    0.457416629856141                    ns
#> 16   0.8043452846734           3    0.480033356696624                    ns
#> 
# }