Diversity Indices

Compute the following diversity indices to compare the phenotypic diversity for qualitative traits between entire collection (EC) and core set (CS).

Simpson's Index (\(D\)) (Simpson 1949)

Nei's Diversity Index or Nei's Variation Index or Simpson's Index of Diversity or Gini's Diversity Index or Gini-Simpson Index (\(H\)) (Gini 1912, 1912; Nei 1973)

Maximum Nei's Diversity/Variation Index or Maximum Simpson's Index of Diversity (\(H_{max}\)) (Hennink and Zeven 1990)

Simpson's Reciprocal Index (\(D_{R}\)) (Hennink and Zeven 1990)

Relative Nei's Diversity/Variation Index or Relative Simpson's Index of Diversity (\(H'\)) (Hennink and Zeven 1990)

Shannon or Shannon-Weaver or Shannon-Weiner Diversity Index (\(I\)) (Shannon and Weaver 1949)

Maximum Shannon-Weaver Diversity Index (\(I_{max}\)) (Hennink and Zeven 1990)

Relative Shannon-Weaver Diversity Index or Shannon Equitability Index (\(I'\)) (Hennink and Zeven 1990)

McIntosh Diversity Index (\(D_{Mc}\)) (McIntosh 1967)

diversity.evaluate.core(data, names, qualitative, selected)

Arguments

data

The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names

Name of column with the individual names as a character string

qualitative

Name of columns with the qualitative traits as a character vector.

selected

Character vector with the names of individuals selected in core collection and present in the names column.

Value

A data frame with the following columns.

Trait

The qualitative trait.

EC_No.Classes

The number of classes in the trait for EC.

EC_D

The Simpson's Index (\(D\)) for EC.

EC_H

Nei's Diversity Index (\(H\)) for EC.

EC_H.max

Maximum Nei's Diversity Index (\(H_{max}\)) for EC.

EC_D.inv

Simpson's Reciprocal Index (\(D_{R}\)) for EC.

EC_H.rel

Relative Nei's Diversity Index (\(H'\)) for EC.

EC_I

Shannon-Weaver Diversity Index (\(I\)) for EC.

EC_I.max

Maximum Shannon-Weaver Diversity Index (\(I_{max}\)) for EC.

EC_I.rel

Relative Shannon-Weaver Diversity Index (\(I'\)) for EC.

EC_D.Mc

McIntosh Diversity Index (\(D_{Mc}\)) for EC.

CS_No.Classes

The number of classes in the trait for CS.

CS_D

The Simpson's Index (\(D\)) for CS.

CS_H

Nei's Diversity Index (\(H\)) for CS.

CS_H.max

Maximum Nei's Diversity Index (\(H_{max}\)) for CS.

CS_D.inv

Simpson's Reciprocal Index (\(D_{R}\)) for CS.

CS_H.rel

Relative Nei's Diversity Index (\(H'\)) for CS.

CS_I

Shannon-Weaver Diversity Index (\(I\)) for CS.

CS_I.max

Maximum Shannon-Weaver Diversity Index (\(I_{max}\)) for CS.

CS_I.rel

Relative Shannon-Weaver Diversity Index (\(I'\)) for CS.

CS_D.Mc

McIntosh Diversity Index (\(D_{Mc}\)) for CS.

.

Details

Simpson's index (\(D\)) which estimates the probability that two accessions randomly selected will belong to the same phenotypic class of a trait, is computed as follows (Simpson 1949) .

\[D = \sum_{i = 1}^{k}p_{i}^{2}\]

Where, \(p_{i}\) denotes the proportion/fraction/frequency of accessions in the \(i\)th phenotypic class for a trait and \(k\) is the number of phenotypic classes for the trait.

The value of \(D\) can range from 0 to 1 with 0 representing maximum diversity and 1, no diversity.

\(D\) is subtracted from 1 to give Nei's diversity index or Nei's variation index or Simpson's index of diversity or Gini's diversity index or Gini-Simpson index \(H\). Greater the value of \(H\), greater the diversity (Gini 1912, 1912; Nei 1973) with a range from 0 to 1.

\[H = 1 - D\]

The maximum value of \(H\), \(H_{max}\) occurs when accessions are uniformly distributed across the phenotypic classes and is computed as follows (Hennink and Zeven 1990) .

\[H_{max} = 1 - \frac{1}{k}\]

Reciprocal of \(D\) gives the Simpson's reciprocal index (\(D_{R}\)) (Hennink and Zeven 1990) and can range from 1 to \(k\).

\[D_{R} = \frac{1}{D}\]

Relative Nei's diversity/variation index or Relative Simpson's index of diversity (\(H'\)) (Hennink and Zeven 1990) is defined as follows.

\[H' = \frac{H}{H_{max}}\]

An index of information \(I\), was described by Shannon and Weaver (1949) as follows.

\[I = -\sum_{i=1}^{k}p_{i} \log_{2}(p_{i})\]

\(I\) is described as Shannon or Shannon-Weaver or Shannon-Weiner diversity index in literature.

Alternatively, \(I\) is also computed using natural logarithm instead of logarithm to base 2.

\[I = -\sum_{i=1}^{k}p_{i} \ln(p_{i})\]

The maximum value of \(I\) (\(I_{max}\)) is \(\ln(k)\). This value occurs when each phenotypic class for a trait has the same proportion of accessions.

\[I_{max} = \log_{2}(k)\] OR \[I_{max} = \ln(k)\]

The relative Shannon-Weaver diversity index or Shannon equitability index (\(I'\)) is the Shannon diversity index (\(I\)) divided by the maximum diversity (\(I_{max}\)).

\[I' = \frac{I}{I_{max}}\]

A similar index of diversity was described by McIntosh (1967) as follows (\(D_{Mc}\)).

\[D_{Mc} = \frac{N - \sqrt{\sum_{i=1}^{k}n_{i}^2}}{N - \sqrt{N}}\]

Where, \(n_{i}\) denotes the number of accessions in the \(i\)th phenotypic class for a trait and \(N\) is the total number of accessions so that \(p_{i} = {n_{i}}/{N}\).

References

Gini C (1912). Variabilita e Mutabilita. Contributo allo Studio delle Distribuzioni e delle Relazioni Statistiche. [Fasc. I.]. Tipogr. di P. Cuppini, Bologna.

Gini C (1912). “Variabilita e mutabilita.” In Pizetti E, Salvemini T (eds.), Memorie di Metodologica Statistica. Liberia Eredi Virgilio Veschi, Roma, Italy.

Hennink S, Zeven AC (1990). “The interpretation of Nei and Shannon-Weaver within population variation indices.” Euphytica, 51(3), 235--240.

McIntosh RP (1967). “An index of diversity and the relation of certain concepts to diversity.” Ecology, 48(3), 392--404.

Nei M (1973). “Analysis of gene diversity in subdivided populations.” Proceedings of the National Academy of Sciences, 70(12), 3321--3323.

Shannon CE, Weaver W (1949). The Mathematical Theory of Communication, number v. 2 in The Mathematical Theory of Communication. University of Illinois Press.

Simpson EH (1949). “Measurement of diversity.” Nature, 163(4148), 688--688.

See also

Examples

data("cassava_CC")
data("cassava_EC")

ec <- cbind(genotypes = rownames(cassava_EC), cassava_EC)
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL

core <- rownames(cassava_CC)

quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

ec[, qual] <- lapply(ec[, qual],
                     function(x) factor(as.factor(x)))

diversity.evaluate.core(data = ec, names = "genotypes",
                        qualitative = qual, selected = core)
#>     Trait EC_No.Classes CS_No.Classes      EC_D      CS_D      EC_H      CS_H
#> 1    CUAL             5             5 0.3792315 0.3784722 0.6207685 0.6215278
#> 2    LNGS             3             3 0.4224002 0.3877551 0.5775998 0.6122449
#> 3    PTLC             5             5 0.4773980 0.4213435 0.5226020 0.5786565
#> 4    DSTA             5             5 0.3514001 0.3035006 0.6485999 0.6964994
#> 5    LFRT             5             5 0.4489353 0.4265873 0.5510647 0.5734127
#> 6   LBTEF             6             6 0.2239712 0.2005385 0.7760288 0.7994615
#> 7    CBTR             3             3 0.5187428 0.4872449 0.4812572 0.5127551
#> 8    NMLB            10            10 0.2067037 0.1980584 0.7932963 0.8019416
#> 9    ANGB             4             4 0.3505164 0.3421202 0.6494836 0.6578798
#> 10 CUAL9M             5             5 0.3090735 0.2880527 0.6909265 0.7119473
#> 11  LVC9M             5             5 0.4252995 0.3886763 0.5747005 0.6113237
#> 12 TNPR9M             5             5 0.2471656 0.2218679 0.7528344 0.7781321
#> 13   PL9M             3             3 0.4986353 0.5002834 0.5013647 0.4997166
#> 14   STRP             4             4 0.3221898 0.3118622 0.6778102 0.6881378
#> 15   STRC             2             2 0.5103256 0.5181406 0.4896744 0.4818594
#> 16   PSTR             3             3 0.5553068 0.5555556 0.4446932 0.4444444
#>     EC_H.max  CS_H.max EC_D.inv CS_D.inv  EC_H.rel  CS_H.rel      EC_I
#> 1  0.8000000 0.8000000 2.636912 2.642202 0.7759606 0.7769097 1.6011803
#> 2  0.6666667 0.6666667 2.367423 2.578947 0.8663998 0.9183673 1.3463809
#> 3  0.8000000 0.8000000 2.094688 2.373360 0.6532525 0.7233206 1.2975799
#> 4  0.8000000 0.8000000 2.845759 3.294887 0.8107499 0.8706243 1.7521779
#> 5  0.8000000 0.8000000 2.227492 2.344186 0.6888308 0.7167659 1.3202043
#> 6  0.8333333 0.8333333 4.464860 4.986572 0.9312346 0.9593537 2.2610792
#> 7  0.6666667 0.6666667 1.927737 2.052356 0.7218857 0.7691327 1.0494076
#> 8  0.9000000 0.9000000 4.837842 5.049016 0.8814403 0.8910462 2.4604366
#> 9  0.7500000 0.7500000 2.852934 2.922949 0.8659782 0.8771731 1.6318128
#> 10 0.8000000 0.8000000 3.235476 3.471587 0.8636581 0.8899341 1.8189573
#> 11 0.8000000 0.8000000 2.351284 2.572835 0.7183757 0.7641546 1.4578126
#> 12 0.8000000 0.8000000 4.045871 4.507186 0.9410430 0.9726651 2.1593638
#> 13 0.6666667 0.6666667 2.005474 1.998867 0.7520470 0.7495748 1.0239323
#> 14 0.7500000 0.7500000 3.103760 3.206544 0.9037469 0.9175170 1.7047164
#> 15 0.5000000 0.5000000 1.959533 1.929978 0.9793487 0.9637188 0.9850516
#> 16 0.6666667 0.6666667 1.800806 1.800000 0.6670399 0.6666667 0.9329026
#>         CS_I  EC_I.max  CS_I.max  EC_I.rel  CS_I.rel   EC_D.Mc   CS_D.Mc
#> 1        NaN 1.6094379 1.6094379 0.9948692       NaN 0.3937780 0.4169689
#> 2  1.4488156 1.0986123 1.0986123 1.2255287 1.3187688 0.3588207 0.4088431
#> 3  1.6068519 1.6094379 1.6094379 0.8062317 0.9983932 0.3167795 0.3802252
#> 4  1.9465251 1.6094379 1.6094379 1.0886893 1.2094440 0.4173809 0.4866359
#> 5        NaN 1.6094379 1.6094379 0.8202891       NaN 0.3382154 0.3758619
#> 6  2.3831722 1.7917595 1.7917595 1.2619323 1.3300738 0.5399006 0.5983483
#> 7  1.1233819 1.0986123 1.0986123 0.9552120 1.0225463 0.2867496 0.3272160
#> 8        NaN 2.3025851 2.3025851 1.0685540       NaN 0.5589746 0.6013583
#> 9  1.6970011 1.3862944 1.3862944 1.1771041 1.2241276 0.4181453 0.4497918
#> 10 1.9268805 1.6094379 1.6094379 1.1301817 1.1972381 0.4551475 0.5020268
#> 11 1.5970207 1.6094379 1.6094379 0.9057899 0.9922847 0.3565384 0.4080420
#> 12 2.2473016 1.6094379 1.6094379 1.3416882 1.3963270 0.5154021 0.5731943
#> 13       NaN 1.0986123 1.0986123 0.9320234       NaN 0.3011986 0.3171624
#> 14 1.7492095 1.3862944 1.3862944 1.2296929 1.2617880 0.4431820 0.4784684
#> 15 0.9736681 0.6931472 0.6931472 1.4211290 1.4047061 0.2927634 0.3036037
#> 16       NaN 1.0986123 1.0986123 0.8491646       NaN 0.2611754 0.2759327