Interquartile Range

Compute the Interquartile Range (IQR) (Upton and Cook 1996) to compare quantitative traits of the entire collection (EC) and core set (CS).

iqr.evaluate.core(data, names, quantitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

quantitative
Name of columns with the quantitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

A data frame with the IQR values of the EC and CS for the traits specified as quantitative.

References

Upton G, Cook I (1996). “General summary statistics.” In Understanding statistics. Oxford University Press.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

iqr.evaluate.core(data = ec, names = "genotypes",
                  quantitative = quant, selected = core)

#>    Trait  EC_IQR  CS_IQR
#> X2    X2 3.42580 3.34580
#> X3    X3 1.10640 1.25080
#> X4    X4 0.02452 0.02955
#> X5    X5 1.22000 1.24000
#> X8    X8 0.80920 1.34680