Distance Between Probability Distributions

Compute Kullback-Leibler (Kullback and Leibler 1951) , Kolmogorov-Smirnov (Kolmogorov 1933; Smirnov 1948) and Anderson-Darling distances (Anderson and Darling 1952) between the probability distributions of collection (EC) and core set (CS) for quantitative traits.

pdfdist.evaluate.core(data, names, quantitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

quantitative
Name of columns with the quantitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

A data frame with the following columns.

Trait

The quantitative trait.

KL_Distance

The Kullback-Leibler distance (Kullback and Leibler 1951) between EC and CS.

KS_Distance

The Kolmogorov-Smirnov distance (Kolmogorov 1933; Smirnov 1948) between EC and CS.

KS_pvalue

The p value of the Kolmogorov-Smirnov distance.

AD_Distance

Anderson-Darling distance (Anderson and Darling 1952) between EC and CS.

AD_pvalue

The p value of the Anderson-Darling distance.

KS_significance

The significance of the Kolmogorov-Smirnov distance (*: p ≤ 0.01; **: p ≤ 0.05; ns: p > 0.05).

AD_pvalue

The significance of the Anderson-Darling distance (*: p ≤ 0.01; **: p ≤ 0.05; ns: p > 0.05).

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

pdfdist.evaluate.core(data = ec, names = "genotypes",
                      quantitative = quant, selected = core)

#> Warning: p-value will be approximate in the presence of ties

#> Warning: p-value will be approximate in the presence of ties

#> Warning: p-value will be approximate in the presence of ties

#> Warning: p-value will be approximate in the presence of ties

#> Warning: p-value will be approximate in the presence of ties

#>   Trait KL_Distance KS_Distance  KS_pvalue AD_Distance  AD_pvalue
#> 1    X2  0.10477307  0.03352160 0.98783942     0.30292 0.93604000
#> 2    X3  0.04422509  0.06656674 0.40549135     1.48630 0.17966000
#> 3    X4  0.02861614  0.07531542 0.26159518     2.54120 0.04728700
#> 4    X5  0.08421608  0.08809463 0.12406512     3.01030 0.02711600
#> 5    X8  0.07322450  0.10720163 0.03260334     5.94860 0.00099556
#>   KS_significance AD_significance
#> 1              ns              ns
#> 2              ns              ns
#> 3              ns               *
#> 4              ns               *
#> 5               *              **