Percentage Difference of Means and Variances

Compute the following differences between the entire collection (EC) and core set (CS).

Percentage of significant differences of mean (MD\%_HuMD\\\%_HuMD%_Hu) hu_methods_2000EvaluateCore

Percentage of significant differences of variance (VD\%_HuVD\\\%_HuVD%_Hu) hu_methods_2000EvaluateCore

Average of absolute differences between means (MD\%_KimMD\\\%_KimMD%_Kim) kim_PowerCore_2007EvaluateCore

Average of absolute differences between variances (VD\%_KimVD\\\%_KimVD%_Kim) kim_PowerCore_2007EvaluateCore

Percentage difference between the mean squared Euclidean distance among accessions (dD\%dD\\\%dD%) studnicki_comparing_2013EvaluateCore

Percentage of range ratios smaller than 0.70 (S_RR_0.7) diwan_methods_1995EvaluateCore

percentdiff.evaluate.core(
  data,
  names,
  quantitative,
  selected,
  alpha = 0.05,
  rr.crit = 0.7
)

Arguments

data

The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names

Name of column with the individual names as a character string.

quantitative

Name of columns with the quantitative traits as a character vector.

selected

Character vector with the names of individuals selected in core collection and present in the names column.

alpha

Type I error probability (Significance level) of difference.

rr.crit

The critical value of range ratio considered to be acceptable for a representative CS. The default value is 0.7.

Value

A data frame with the values of MD\%_HuMD\\\%_HuMD%_Hu, VD\%_HuVD\\\%_HuVD%_Hu, MD\%_KimMD\\\%_KimMD%_Kim, VD\%_KimVD\\\%_KimVD%_Kim and dD\%dD\\\%dD%.

Details

The differences are computed as follows.

MD\%_Hu = ( S_tn ) 100MD\\\%_Hu = ( S_tn ) 100MD%_Hu = ( S_tn ) 100

Where, S_t is the number of traits with a significant difference between the means of the EC and the CS and n is the total number of traits. A representative core should have MD\%_HuMD\\\%_HuMD%_Hu < 20 % and CR > 80 % hu_methods_2000EvaluateCore.

VD\%_Hu = ( S_Fn ) 100VD\\\%_Hu = ( S_Fn ) 100VD%_Hu = ( S_Fn ) 100

Where, S_F is the number of traits with a significant difference between the variances of the EC and the CS and n is the total number of traits. Larger VD\%_HuVD\\\%_HuVD%_Hu value indicates a more diverse core set.

MD\%_Kim = ( 1n_i=1^n | M_EC_i-M_CS_i |M_CS_i ) 100MD\\\%_Kim = ( 1n_i=1^n | M_EC_i-M_CS_i |M_CS_i ) 100MD%_Kim = ( 1n_i=1^n | M_EC_i-M_CS_i |M_CS_i ) 100

Where, M_EC_i is the mean of the EC for the ith trait, M_CS_i is the mean of the CS for the ith trait and n is the total number of traits.

VD\%_Kim = ( 1n_i=1^n | V_EC_i-V_CS_i |V_CS_i ) 100VD\\\%_Kim = ( 1n_i=1^n | V_EC_i-V_CS_i |V_CS_i ) 100VD%_Kim = ( 1n_i=1^n | V_EC_i-V_CS_i |V_CS_i ) 100

Where, V_EC_i is the variance of the EC for the ith trait, V_CS_i is the variance of the CS for the ith trait and n is the total number of traits.

dD\% = d_CS-d_ECd_EC 100dD\\\% = d_CS-d_ECd_EC 100dD\ d_CS-d_ECd_EC 100

Where, d_CS is the mean squared Euclidean distance among accessions in the CS and d_EC is the mean squared Euclidean distance among accessions in the EC.

Percentage of range ratios smaller than 0.70 diwan_methods_1995EvaluateCore is computed as follows.

RR\%_0.7 = ( S_RR_0.7n ) 100RR\\\%_0.7 = ( S_RR_0.7n ) 100RR%_0.7 = ( S_RR_0.7n ) 100

Where, S_RR_0.7 is the number of traits with a range ratio smaller than 0.7 (R_CS_iR_EC_i < 0.7) R_CS_i is the range of the ith trait in the CS, R_EC_i is the range of the ith trait in the EC and n is the total number of traits.

References

See also

Examples

data("cassava_CC")
data("cassava_EC")

ec <- cbind(genotypes = rownames(cassava_EC), cassava_EC)
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL

core <- rownames(cassava_CC)

quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

ec[, qual] <- lapply(ec[, qual],
                     function(x) factor(as.factor(x)))

percentdiff.evaluate.core(data = ec, names = "genotypes",
                          quantitative = quant, selected = core)
#>   MDPercent_Hu VDPercent_Hu MDPercent_Kim VDPercent_Kim DDPercent RR
#> 1           50           80      13.02737      41.64331   18.2052 20