Percentage Difference of Means and Variances

Compute the following differences between the entire collection (EC) and core set (CS).

Percentage of significant differences of mean (\(MD\%_{Hu}\)) (Hu et al. 2000)

Percentage of significant differences of variance (\(VD\%_{Hu}\)) (Hu et al. 2000)

Average of absolute differences between means (\(MD\%_{Kim}\)) (Kim et al. 2007)

Average of absolute differences between variances (\(VD\%_{Kim}\)) (Kim et al. 2007)

Percentage difference between the mean squared Euclidean distance among accessions (\(\overline{d}D\%\)) (Studnicki et al. 2013)

percentdiff.evaluate.core(data, names, quantitative, selected, alpha = 0.05)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

quantitative
Name of columns with the quantitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

alpha
Type I error probability (Significance level) of difference.

Value

A data frame with the values of \(MD\%_{Hu}\), \(VD\%_{Hu}\), \(MD\%_{Kim}\), \(VD\%_{Kim}\) and \(\overline{d}D\%\).

Details

The differences are computed as follows.

\[MD\%_{Hu} = \left ( \frac{S_{t}}{n} \right ) \times 100\]

Where, \(V_{EC_{i}}\) is the variance of the EC for the \(i\)th trait, \(V_{CS_{i}}\) is the variance of the CS for the \(i\)th trait and \(n\) is the total number of traits.

\[\overline{d}D\% = \frac{\overline{d}_{CS}-\overline{d}_{EC}}{\overline{d}_{EC}} \times 100\]

Where, \(\overline{d}_{CS}\) is the mean squared Euclidean distance among accessions in the CS and \(\overline{d}_{EC}\) is the mean squared Euclidean distance among accessions in the EC.

References

Hu J, Zhu J, Xu HM (2000). “Methods of constructing core collections by stepwise clustering with three sampling strategies based on the genotypic values of crops.” Theoretical and Applied Genetics, 101(1), 264--268.

Kim K, Chung H, Cho G, Ma K, Chandrabalan D, Gwag J, Kim T, Cho E, Park Y (2007). “PowerCore: A program applying the advanced M strategy with a heuristic search for establishing core sets.” Bioinformatics, 23(16), 2155--2162.

Studnicki M, Madry W, Schmidt J (2013). “Comparing the efficiency of sampling strategies to establish a representative in the phenotypic-based genetic diversity core collection of orchardgrass (Dactylis glomerata L.).” Czech Journal of Genetics and Plant Breeding, 49(1), 36--47.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

percentdiff.evaluate.core(data = ec, names = "genotypes",
                          quantitative = quant, selected = core)

#>   MDPercent_Hu VDPercent_Hu MDPercent_Kim VDPercent_Kim DDPercent
#> 1           40           60      1.479138      27.37944  11.66411