Student-Newman-Keuls Test

Test difference between means of entire collection (EC) and core set (CS) for quantitative traits by Newman-Keuls or Student-Newman-Keuls test (Newman 1939; Keuls 1952) .

snk.evaluate.core(data, names, quantitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

quantitative
Name of columns with the quantitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

A data frame with the following components.

Trait

The quantitative trait.

EC_Min

The minimum value of the trait in EC.

EC_Max

The maximum value of the trait in EC.

EC_Mean

The mean value of the trait in EC.

EC_SE

The standard error of the trait in EC.

CS_Min

The minimum value of the trait in CS.

CS_Max

The maximum value of the trait in CS.

CS_Mean

The mean value of the trait in CS.

CS_SE

The standard error of the trait in CS.

SNK_pvalue

The p value of the Student-Newman-Keuls test for equality of means of EC and CS.

SNK_significance

The significance of the Student-Newman-Keuls test for equality of means of EC and CS.

References

Keuls M (1952). “The use of the ,,studentized range" in connection with an analysis of variance.” Euphytica, 1(2), 112--122.

Newman D (1939). “The distribution of range in samples from a normal population, expressed in terms of an independent estimate of standard deviation.” Biometrika, 31(1-2), 20--30.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

snk.evaluate.core(data = ec, names = "genotypes",
                  quantitative = quant, selected = core)

#>   Trait     EC_Min     EC_Max    EC_Mean       EC_SE     CS_Min     CS_Max
#> 1    X2  6.8701433 26.9501433 15.1037469 0.056799683  8.2701433 26.9501433
#> 2    X3 20.5105583 27.5849583 22.9311580 0.018661701 20.6308583 27.5849583
#> 3    X4  0.6881624  0.8673024  0.7958625 0.000406744  0.7485624  0.8519724
#> 4    X5 -1.3707739  5.6692261  2.6531327 0.019753121 -1.0707739  5.0692261
#> 5    X8 55.8195689 74.3669689 67.0581697 0.018214460 58.5131689 74.3669689
#>      CS_Mean       CS_SE SNK_pvalue SNK_significance
#> 1 15.1478967 0.195862983     0.8163               ns
#> 2 22.9912013 0.071270614     0.3420               ns
#> 3  0.7984705 0.001450459     0.0563               ns
#> 4  2.4959266 0.066030538     0.0171                *
#> 5 67.2046329 0.088740924     0.0223                *