Student's t Test

Test difference between means of entire collection (EC) and core set (CS) for quantitative traits by Student's t test (Student 1908) .

ttest.evaluate.core(data, names, quantitative, selected)

Arguments

data

The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names

Name of column with the individual names as a character string

quantitative

Name of columns with the quantitative traits as a character vector.

selected

Character vector with the names of individuals selected in core collection and present in the names column.

Value

Trait

The quantitative trait.

EC_Min

The minimum value of the trait in EC.

EC_Max

The maximum value of the trait in EC.

EC_Mean

The mean value of the trait in EC.

EC_SE

The standard error of the trait in EC.

CS_Min

The minimum value of the trait in CS.

CS_Max

The maximum value of the trait in CS.

CS_Mean

The mean value of the trait in CS.

CS_SE

The standard error of the trait in CS.

ttest_pvalue

The p value of the Student's t test for equality of means of EC and CS.

ttest_significance

The significance of the Student's t test for equality of means of EC and CS.

References

Student (1908). “The probable error of a mean.” Biometrika, 6(1), 1--25.

See also

Examples

data("cassava_CC")
data("cassava_EC")

ec <- cbind(genotypes = rownames(cassava_EC), cassava_EC)
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL

core <- rownames(cassava_CC)

quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
           "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

ec[, qual] <- lapply(ec[, qual],
                     function(x) factor(as.factor(x)))

ttest.evaluate.core(data = ec, names = "genotypes",
                    quantitative = quant, selected = core)
#>    Trait     EC_Min    EC_Max       EC_Mean      EC_SE     CS_Min    CS_Max
#> 1   NMSR -1.3948656  5.630140 -6.330344e-17 0.02436851 -1.3948656  5.630140
#> 2   TTRN -1.8861153  5.179145 -3.572527e-17 0.02436851 -1.8861153  5.179145
#> 3  TFWSR -1.1977304  7.626990  1.696184e-17 0.02436851 -1.1536068  7.185754
#> 4   TTRW -1.1454311 11.045474 -8.014340e-19 0.02436851 -1.0850801 11.045474
#> 5  TFWSS -1.1443168  5.949272 -7.686375e-18 0.02436851 -1.1443168  5.949272
#> 6   TTSW -1.1725141  9.795915  6.181556e-17 0.02436851 -1.1425457  9.795915
#> 7   TTPW -1.2231612  6.909430 -9.405196e-17 0.02436851 -1.2231612  6.909430
#> 8   AVPW -1.2146948  8.537471 -2.728275e-18 0.02436851 -1.2146948  8.537471
#> 9   ARSR -0.8242221  7.160380 -3.455680e-17 0.02436851 -0.8242221  2.724490
#> 10  SRDM -7.4054956  2.211252  7.042075e-16 0.02436851 -3.1534628  2.052297
#>         CS_Mean      CS_SE ttest_pvalue ttest_significance
#> 1  -0.107877169 0.08214456  0.209505833                 ns
#> 2   0.040164914 0.09179951  0.672855738                 ns
#> 3   0.202799940 0.10360518  0.058266745                 ns
#> 4   0.434448321 0.13321838  0.001585008                 **
#> 5   0.136638483 0.09483542  0.164507433                 ns
#> 6   0.340433536 0.11894631  0.005599998                 **
#> 7   0.176178665 0.09931291  0.086559633                 ns
#> 8   0.416683438 0.12519928  0.001302429                 **
#> 9  -0.069064618 0.06726480  0.335457319                 ns
#> 10 -0.008079704 0.06536988  0.907907091                 ns