Wilcoxon Rank Sum Test

Compare the medians of quantitative traits between entire collection (EC) and core set (CS) by Wilcoxon rank sum test or Mann-Whitney-Wilcoxon test or Mann-Whitney U test (Wilcoxon 1945; Mann and Whitney 1947) .

wilcox.evaluate.core(data, names, quantitative, selected)

Arguments

data
The data as a data frame object. The data frame should possess one row per individual and columns with the individual names and multiple trait/character data.

names
Name of column with the individual names as a character string

quantitative
Name of columns with the quantitative traits as a character vector.

selected
Character vector with the names of individuals selected in core collection and present in the names column.

Value

Trait

The quantitative trait.

EC_Med

The median value of the trait in EC.

CS_Med

The median value of the trait in CS.

Wilcox_pvalue

The p value of the Wilcoxon test for equality of medians of EC and CS.

Wilcox_significance

The significance of the Wilcoxon test for equality of medians of EC and CS.

References

Mann HB, Whitney DR (1947). “On a test of whether one of two random variables is stochastically larger than the other.” The Annals of Mathematical Statistics, 18(1), 50--60.

Wilcoxon F (1945). “Individual comparisons by ranking methods.” Biometrics Bulletin, 1(6), 80.

See also

Examples

####################################
# Use data from R package ccChooser
####################################

library(ccChooser)
data("dactylis_CC")
data("dactylis_EC")

ec <- cbind(genotypes = rownames(dactylis_EC), dactylis_EC[, -1])
ec$genotypes <- as.character(ec$genotypes)
rownames(ec) <- NULL
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) cut(x, breaks = 4))
ec[, c("X1", "X6", "X7")] <- lapply(ec[, c("X1", "X6", "X7")],
                                    function(x) factor(as.numeric(x)))
head(ec)

#>    genotypes X1       X2       X3        X4       X5 X6 X7       X8
#> 1       ASTA  2 16.67014 23.51726 0.8068824 2.029226  2  3 67.36377
#> 2    BARAULA  3 14.52438 23.32746 0.7897424 2.949226  3  4 67.93947
#> 3     BRAULA  3 15.15014 22.84370 0.7818024 3.369226  2  3 67.40187
#> 4 BRUDZYNSKA  2 14.45010 22.79556 0.7910424 3.269226  2  2 67.40187
#> 5    D120001  2 13.85014 21.44856 0.7707224 3.269226  2  3 67.40187
#> 6    D120002  2 15.15014 22.50696 0.7928824 3.169226  2  2 67.40187

core <- rownames(dactylis_CC)

quant <- c("X2", "X3", "X4", "X5", "X8")
qual <- c("X1", "X6", "X7")

####################################
# EvaluateCore
####################################

wilcox.evaluate.core(data = ec, names = "genotypes",
                     quantitative = quant, selected = core)

#>   Trait     EC_Med     CS_Med Wilcox_pvalue Wilcox_significance
#> 1    X2 15.0243433 14.8701433    0.91245698                  ns
#> 2    X3 22.8677483 22.9158983    0.67483316                  ns
#> 3    X4  0.7965774  0.7965774    0.11336957                  ns
#> 4    X5  2.6692261  2.5492261    0.01817080                   *
#> 5    X8 66.8630689 67.3637689    0.01155352                   *